+++ featuretools/__init__.py
@@ -5,7 +5,8 @@ from . import variable_types
 from .entityset.api import *
 from . import primitives
 from .synthesis.api import *
-from .primitives import Feature, list_primitives
+from .primitive_utils import Feature
+from .primitives import  list_primitives
 from .computational_backends.api import *
 from . import tests
 from .utils.pickle_utils import *
