+++ featuretools/primitive_utils/api.py
@@ -1,0 +1,7 @@
+# flake8: noqa
+from .aggregation_primitive_base import (
+    AggregationPrimitive,
+    make_agg_primitive
+)
+from .primitive_base import PrimitiveBase, IdentityFeature, DirectFeature, Feature
+from .transform_primitive_base import TransformPrimitive, make_trans_primitive
