+++ featuretools/primitive_utils/install.py
@@ -1,0 +1,5 @@
+def install_primitives(directory):
+    # iterate over files
+    # if a file has a primitive in it copy to installed
+    # track name of primitive
+    # print that it's been installed
