+++ featuretools/primitive_utils/primitive_base.py
@@ -150,7 +150,7 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.equal_to`
         """
-        from .binary_transform import Equals
+        from featuretools.primitives.binary_transform import Equals
         return Equals(self, other_feature_or_val)
 
     def __ne__(self, other_feature_or_val):
@@ -159,7 +159,7 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.not_equal_to`
         """
-        from .binary_transform import NotEquals
+        from featuretools.primitives.binary_transform import NotEquals
         return NotEquals(self, other_feature_or_val)
 
     def __gt__(self, other_feature_or_val):
@@ -168,7 +168,7 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.GT`
         """
-        from .binary_transform import GreaterThan
+        from featuretools.primitives.binary_transform import GreaterThan
         return GreaterThan(self, other_feature_or_val)
 
     def __ge__(self, other_feature_or_val):
@@ -177,7 +177,7 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.greater_than_equal_to`
         """
-        from .binary_transform import GreaterThanEqualTo
+        from featuretools.primitives.binary_transform import GreaterThanEqualTo
         return GreaterThanEqualTo(self, other_feature_or_val)
 
     def __lt__(self, other_feature_or_val):
@@ -186,7 +186,7 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.less_than`
         """
-        from .binary_transform import LessThan
+        from featuretools.primitives.binary_transform import LessThan
         return LessThan(self, other_feature_or_val)
 
     def __le__(self, other_feature_or_val):
@@ -195,16 +195,16 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.less_than_equal_to`
         """
-        from .binary_transform import LessThanEqualTo
+        from featuretools.primitives.binary_transform import LessThanEqualTo
         return LessThanEqualTo(self, other_feature_or_val)
 
     def __add__(self, other_feature_or_val):
         """Add other_feature_or_val"""
-        from .binary_transform import Add
+        from featuretools.primitives.binary_transform import Add
         return Add(self, other_feature_or_val)
 
     def __radd__(self, other):
-        from .binary_transform import Add
+        from featuretools.primitives.binary_transform import Add
         return Add(other, self)
 
     def __sub__(self, other_feature_or_val):
@@ -213,11 +213,11 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.subtract`
         """
-        from .binary_transform import Subtract
+        from featuretools.primitives.binary_transform import Subtract
         return Subtract(self, other_feature_or_val)
 
     def __rsub__(self, other):
-        from .binary_transform import Subtract
+        from featuretools.primitives.binary_transform import Subtract
         return Subtract(other, self)
 
     def __div__(self, other_feature_or_val):
@@ -226,18 +226,18 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.divide`
         """
-        from .binary_transform import Divide
+        from featuretools.primitives.binary_transform import Divide
         return Divide(self, other_feature_or_val)
 
     def __truediv__(self, other_feature_or_val):
         return self.__div__(other_feature_or_val)
 
     def __rtruediv__(self, other_feature_or_val):
-        from .binary_transform import Divide
+        from featuretools.primitives.binary_transform import Divide
         return Divide(other_feature_or_val, self)
 
     def __rdiv__(self, other_feature_or_val):
-        from .binary_transform import Divide
+        from featuretools.primitives.binary_transform import Divide
         return Divide(other_feature_or_val, self)
 
     def __mul__(self, other_feature_or_val):
@@ -246,11 +246,11 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.multiply`
         """
-        from .binary_transform import Multiply
+        from featuretools.primitives.binary_transform import Multiply
         return Multiply(self, other_feature_or_val)
 
     def __rmul__(self, other):
-        from .binary_transform import Multiply
+        from featuretools.primitives.binary_transform import Multiply
         return Multiply(other, self)
 
     def __mod__(self, other_feature_or_val):
@@ -259,64 +259,64 @@ class PrimitiveBase(object):
         See also:
             :meth:`PrimitiveBase.modulo`
         """
-        from .binary_transform import Mod
+        from featuretools.primitives.binary_transform import Mod
         return Mod(self, other_feature_or_val)
 
     def __and__(self, other):
         return self.AND(other)
 
     def __rand__(self, other):
-        from .binary_transform import And
+        from featuretools.primitives.binary_transform import And
         return And(other, self)
 
     def __or__(self, other):
         return self.OR(other)
 
     def __ror__(self, other):
-        from .binary_transform import Or
+        from featuretools.primitives.binary_transform import Or
         return Or(other, self)
 
     def __not__(self, other):
         return self.NOT(other)
 
     def __abs__(self):
-        from .transform_primitive import Absolute
+        from featuretools.primitives.transform_primitive import Absolute
         return Absolute(self)
 
     def __neg__(self):
-        from .binary_transform import Negate
+        from featuretools.primitives.binary_transform import Negate
         return Negate(self)
 
     def AND(self, other_feature):
         """Logical AND with other_feature"""
-        from .binary_transform import And
+        from featuretools.primitives.binary_transform import And
         return And(self, other_feature)
 
     def OR(self, other_feature):
         """Logical OR with other_feature"""
-        from .binary_transform import Or
+        from featuretools.primitives.binary_transform import Or
         return Or(self, other_feature)
 
     def NOT(self):
         """Creates inverse of feature"""
-        from .transform_primitive import Not
-        from .binary_transform import Compare
+        from featuretools.primitives.transform_primitive import Not
+        from featuretools.primitives.binary_transform import Compare
         if isinstance(self, Compare):
             return self.invert()
         return Not(self)
 
     def LIKE(self, like_string, case_sensitive=False):
-        from .transform_primitive import Like
+        from featuretools.primitives.transform_primitive import Like
         return Like(self, like_string,
                     case_sensitive=case_sensitive)
 
     def isin(self, list_of_output):
-        from .transform_primitive import IsIn
+        from featuretools.primitives.transform_primitive import IsIn
         return IsIn(self, list_of_outputs=list_of_output)
 
     def is_null(self):
         """Compares feature to null by equality"""
-        from .transform_primitive import IsNull
+        from featuretools.primitives.transform_primitive import IsNull
         return IsNull(self)
 
     def __invert__(self):
@@ -454,14 +454,48 @@ class IdentityFeature(PrimitiveBase):
         return 0
 
 
+class DirectFeature(PrimitiveBase):
+    """Feature for child entity that inherits
+        a feature value from a parent entity"""
+    input_types = [Variable]
+    return_type = None
+
+    def __init__(self, base_feature, child_entity):
+        base_feature = self._check_feature(base_feature)
+        if base_feature.expanding:
+            self.expanding = True
+
+        path = child_entity.entityset.find_forward_path(child_entity.id, base_feature.entity.id)
+        if len(path) > 1:
+            parent_entity_id = path[1].child_entity.id
+            parent_entity = child_entity.entityset[parent_entity_id]
+            parent_feature = DirectFeature(base_feature, parent_entity)
+        else:
+            parent_feature = base_feature
+
+        self.parent_entity = parent_feature.entity
+        self._variable_type = parent_feature.variable_type
+        super(DirectFeature, self).__init__(child_entity, [parent_feature])
+
+    @property
+    def default_value(self):
+        return self.base_features[0].default_value
+
+    @property
+    def variable(self):
+        return getattr(self.base_features[0], 'variable', None)
+
+    def generate_name(self):
+        return u"%s.%s" % (self.parent_entity.id,
+                           self.base_features[0].get_name())
+
+
 class Feature(PrimitiveBase):
     """
     Alias for IdentityFeature and DirectFeature depending on arguments
     """
 
     def __new__(self, feature_or_var, entity=None):
-        from . import direct_feature
-
         if entity is None:
             assert isinstance(feature_or_var, (Variable))
             return IdentityFeature(feature_or_var)
@@ -472,4 +506,4 @@ class Feature(PrimitiveBase):
         if feature_or_var.entity.id == entity.id:
             return IdentityFeature(entity)
 
-        return direct_feature.DirectFeature(feature_or_var, entity)
+        return DirectFeature(feature_or_var, entity)
