+++ featuretools/primitives/api.py
@@ -1,13 +1,9 @@
 # flake8: noqa
-from .aggregation_primitive_base import *
 from .aggregation_primitives import *
 from .binary_transform import *
 from .cum_transform_feature import *
-from .direct_feature import *
-from .primitive_base import *
 from .transform_primitive import *
 from .utils import (
-    apply_dual_op_from_feat,
     get_aggregation_primitives,
     get_transform_primitives,
     list_primitives
