+++ featuretools/synthesis/deep_feature_synthesis.py
@@ -4,11 +4,10 @@ from collections import defaultdict
 
 import featuretools.primitives.api as ftypes
 from featuretools import variable_types
+from featuretools.primitive_utils import AggregationPrimitive, DirectFeature
 from featuretools.primitives.api import (
-    AggregationPrimitive,
     BinaryFeature,
     Compare,
-    DirectFeature,
     Discrete,
     Equals,
     IdentityFeature,
@@ -28,7 +27,7 @@ class DeepFeatureSynthesis(object):
 
             entityset (EntitySet): Entityset for which to build features.
 
-            agg_primitives (list[str or :class:`.primitives.AggregationPrimitive`], optional):
+            agg_primitives (list[str or :class:`.primitives.`], optional):
                 list of Aggregation Feature types to apply.
 
                 Default: ["sum", "std", "max", "skew", "min", "mean", "count", "percent_true", "num_unique", "mode"]
