+++ featuretools/tests/primitive_tests/test_primitive_base.py
@@ -2,8 +2,8 @@ import pytest
 from pympler.asizeof import asizeof
 
 from ..testing_utils import make_ecommerce_entityset
-
-from featuretools.primitives import Feature, IdentityFeature, Last, Mode, Sum
+from featuretools.primitive_utils import Feature, IdentityFeature
+from featuretools.primitives import Last, Mode, Sum
 from featuretools.variable_types import Datetime, Numeric
